tidy up - leave only good details
