+++ templates/case/popups/good.html
@@ -43,39 +43,6 @@
 			<h2 class="app-popup__heading">ARS</h2>
 			<p class="app-popup__body" data-max-length="200">Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nam eu enim ex. Cras blandit vehicula pharetra. Phasellus commodo nec lacus tristique rutrum. Donec in elementum neque. Nunc interdum ex vitae vulputate dictum. Praesent tincidunt erat elit, lobortis pulvinar leo semper id. Duis sollicitudin aliquam enim quis consequat. Vestibulum viverra at urna sed sodales. Quisque tincidunt placerat elit id euismod. Integer non odio mollis, ultrices odio facilisis, tristique arcu.</p>
 		</div>
-		<div id="pane-destinations" class="app-hidden">
-			<table class="govuk-table">
-				<thead class="govuk-table__head">
-					<tr class="govuk-table__row">
-						<th scope="col" class="govuk-table__header">Country</th>
-						<th scope="col" class="govuk-table__header"></th>
-					</tr>
-				</thead>
-				<tbody class="govuk-table__body">
-					<tr class="govuk-table__row">
-						<th scope="row" class="govuk-table__header">United States</th>
-						<td class="govuk-table__cell">
-							<a href="{% url 'cases:destination' queue.id case.id case.id %}">View</a>
-						</td>
-					</tr>
-					<tr class="govuk-table__row">
-						<th scope="row" class="govuk-table__header">Poland</th>
-						<td class="govuk-table__cell">
-							<a href="#" onclick="return openDestinationModal()">View</a>
-						</td>
-					</tr>
-					<tr class="govuk-table__row">
-						<th scope="row" class="govuk-table__header">Saudi Arabia</th>
-						<td class="govuk-table__cell">
-							<a href="#" onclick="return openDestinationModal()">View</a>
-						</td>
-					</tr>
-				</tbody>
-			</table>
-		</div>
-		<div id="pane-activity" class="app-hidden">
-			{% include 'includes/audit-trail.html' %}
-		</div>
 	</div>
 	<script type="text/javascript">
 		$(".lite-tabs__tab").click(function() {
