+++ templates/case/slices/destinations.html
@@ -27,7 +27,6 @@
 						<th scope="col" class="govuk-table__header">Website</th>
 						<th scope="col" class="govuk-table__header">Documents</th>
 						<th scope="col" class="govuk-table__header">Flags</th>
-						<th scope="col" class="govuk-table__header"></th>
 					</tr>
 				</thead>
 				<tbody class="govuk-table__body">
@@ -56,9 +55,6 @@
 							<td class="govuk-table__cell {% if show_advice %}lite-!-no-border{% endif %}">
 								{% include 'includes/flags.html' with flags=destination.flags list=True %}
 							</td>
-							<td class="govuk-table__cell {% if show_advice %}lite-!-no-border{% endif %}">
-								<a href="#">View</a>
-							</td>
 						</tr>
 						{% if show_advice %}
 							{% include 'includes/advice.html' with id=destination.id %}
@@ -78,7 +74,6 @@
 						<th scope="col" class="govuk-table__header">Country code</th>
 						<th scope="col" class="govuk-table__header">Goods</th>
 						<th scope="col" class="govuk-table__header">Flags</th>
-						<th scope="col" class="govuk-table__header"></th>
 					</tr>
 				</thead>
 				<tbody class="govuk-table__body">
@@ -106,9 +101,6 @@
 									<br>
 									{% include 'includes/flags.html' with flags=destination.flags list=True %}
 								{% endif %}
-							</td>
-							<td class="govuk-table__cell {% if show_advice %}lite-!-no-border{% endif %}">
-								<a onclick="return openGoodModal()">View</a>
 							</td>
 						</tr>
 						{% if show_advice %}
